requirements.txt
ortools
openpyxl
pandas
python-dateutil
jpholiday
python-dotenv
fastapi
uvicorn[standard]
requests
httpx
